Updated Wator and Segregation simulations

See merge request CompSci308_2018Spring/cellsociety_team18!11

+++ src/cell/segregation/OCell.java
@@ -15,7 +15,7 @@ public class OCell extends SegregationSimCell{
 
 	@Override
 	public Cell nextState(HashSet<Cell> neighbors) {
-		int ocells = 0;
+		double ocells = 0;
 		for (Cell cell:neighbors)
 			if (cell instanceof OCell)
 				ocells++;
